sixerr/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.3
gunicorn==19.7.1
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
stripe==1.62.0
whitenoise==3.3.0
taxjar==1.1.2
